// ui/src/generated/models/bucket.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Photos NG API
 * API for managing albums and media in Photos NG application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Bucket
 */
export interface Bucket {
    /**
     * 
     * @type {number}
     * @memberof Bucket
     */
    'year'?: number;
    /**
     * 
     * @type {number}
     * @memberof Bucket
     */
    'month'?: number;
    /**
     * list of media hrefs in this bucket
     * @type {Array<string>}
     * @memberof Bucket
     */
    'media'?: Array<string>;
}