// ui/src/generated/models/get-timeline-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Photos NG API
 * API for managing albums and media in Photos NG application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Bucket } from './bucket';

/**
 * 
 * @export
 * @interface GetTimelineResponse
 */
export interface GetTimelineResponse {
    /**
     * 
     * @type {Array<Bucket>}
     * @memberof GetTimelineResponse
     */
    'buckets': Array<Bucket>;
    /**
     * List of all years that contain photos
     * @type {Array<number>}
     * @memberof GetTimelineResponse
     */
    'years': Array<number>;
    /**
     * Total number of buckets
     * @type {number}
     * @memberof GetTimelineResponse
     */
    'total': number;
    /**
     * Number of buckets returned
     * @type {number}
     * @memberof GetTimelineResponse
     */
    'limit': number;
    /**
     * Number of buckets skipped
     * @type {number}
     * @memberof GetTimelineResponse
     */
    'offset': number;
}